HW06/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cloud-sql-python-connector==1.4.2
cryptography==41.0.5
frozenlist==1.4.0
google-auth==2.23.4
idna==3.4
joblib==1.3.2
multidict==6.0.4
numpy==1.26.1
pandas==2.1.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.22
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
